# Optimize `get_equivalent_smallest_vectors` (also, bugfix for HDF5 in python3)