@atztogo 2017-04-08
Merged to the master branch which invokes its PyPI and conda releases. Thanks for your beautiful hack, which is fast and readable. Thanks also for the detailed explanation. I replaced the line np.linalg.norm(x, axis=1) by np.sqrt(x**2.sum(axis=1)) because older numpy doesn't support axis keyword in np.linalg.norm.