@atztogo 2017-05-05
A follow-up. I moved get_smallest_vectors method from harmonic/dynamical_matrix.py to structure/cells.py, and it is changed to be called only once at the initialization of Primitive class object for the usual phonopy run. These data are stored in each Primitive class object and accessed by primitive.get_smallest_vectors().
https://github.com/atztogo/phonopy/blob/master/phonopy/structure/cells.py#L299

Togo